Separate chart classes and change class structure
