+++ index.html
@@ -9,63 +9,8 @@
 <body>
   <!-- HTML code -->
   <div class="container">
-    <!-- <div id="chart">
-      <h2 class="title">Chart title</h2>
-      <svg class="chart" width="800" height="640" viewBox="0 0 800 640">
-        <g class="axis x-axis">
-          <text x="720" y="620">x축 이름</text>
-          <line x1="40" x2="760" y1="560" y2="560" />
-        </g>
-        <g class="axis y-axis">
-          <text x="0" y="20">y축 이름</text>
-          <line x1="40" x2="40" y1="40" y2="560" />
-        </g>
-        <g class="labels x-axis-labels">
-          <text x="20" y="590">2015</text>
-          <text x="140" y="590">2016</text>
-          <text x="260" y="590">2017</text>
-          <text x="380" y="590">2018</text>
-          <text x="500" y="590">2019</text>
-          <text x="620" y="590">2020</text>
-          <text x="740" y="590">2021</text>
-        </g>
-        <g class="labels y-axis-labels">
-          <text x="20" y="560">0</text>
-          <text x="20" y="460">10</text>
-          <text x="20" y="360">20</text>
-          <text x="20" y="260">30</text>
-          <text x="20" y="160">40</text>
-          <text x="20" y="60">50</text>
-        </g>
-        <g class="grid y-axis-grid">
-          <line x1="40" x2="760" y1="560" y2="560" />
-          <line x1="40" x2="760" y1="460" y2="460" />
-          <line x1="40" x2="760" y1="360" y2="360" />
-          <line x1="40" x2="760" y1="260" y2="260" />
-          <line x1="40" x2="760" y1="160" y2="160" />
-          <line x1="40" x2="760" y1="60" y2="60" />
-        </g>
-        <g class="data">
-          <g class="area">
-            <path fill="rgba(96, 198, 187, 0.5)" d="M40,560 L 40,460 160,260 280,310 400,60 520,560 640,160 760,60 760,560 40,560 Z" />
-          </g>
-          <g class="line">
-            <path fill="none" d="M 40,460 L 160,260 280,310 400,60 520,560 640,160 760,60" />
-          </g>
-          <g class="dots">
-            <circle cx="40" cy="460" r="8" fill="#fff" data-value="10" />
-            <circle cx="160" cy="260" r="8" fill="#fff" data-value="30" />
-            <circle cx="280" cy="310" r="8" fill="#fff" data-value="25" />
-            <circle cx="400" cy="60" r="8" fill="#fff" data-value="50" />
-            <circle cx="520" cy="560" r="8" fill="#fff" data-value="0" />
-            <circle cx="640" cy="160" r="8" fill="#fff" data-value="40" />
-            <circle cx="760" cy="60" r="8" fill="#fff" data-value="50" />
-          </g>
-        </g>
-      </svg>
-    </div> -->
     <div id="dopyo-chart"></div>
-    <button class="add">Add random data</button>
+    <!-- <button class="add">Add random data</button> -->
   </div>
 </body>
 </html>
