+++ src/main.js
@@ -1,12 +1,14 @@
 import './assets/sass/main.sass';
-import Dopyo from './js/charts/dopyo';
+import ChartBasic from './js/charts/chartBasic';
+import chart from './js/charts/chart';
 
 if (module.hot) {
   module.hot.accept()
 }
 
 document.addEventListener('DOMContentLoaded', () => {
-  const dopyo = new Dopyo({
+  const dopyo = chart.createChart({
+    type: 'line',
     size: {
       width: 600,
       height: 400,
@@ -38,29 +40,4 @@ document.addEventListener('DOMContentLoaded', () => {
       },
     }
   });
-  dopyo.init();
-
-
-  // *********************
-  // Test
-  // *********************
-
-  let count = 6;
-  const addBtn = document.querySelector('.add');
-  addBtn.addEventListener("click", () => {
-    dopyo.addData({
-      xAxis: `2019-0${count}`,
-      series: [
-        {
-          name: '라인그래프1',
-          data: Math.floor(Math.random() * (800 - 0)) + 0
-        },
-        {
-          name: '라인그래프2',
-          data: Math.floor(Math.random() * (800 - 0)) + 0
-        },
-      ],
-    });
-    count++
-  })
 })
